+++ inst/ShinyStan/server_files/pages/diagnose/ppcheck/server/y_vs_avg_rep.R
@@ -16,12 +16,12 @@
 
 pp_y_vs_avg_rep <- reactive({
   pp_tests()
-  y <- get(input$y_name)
-  y_rep <- y_rep()
+  y <- get_y()
+  yrep <- get_yrep()
   zoom <- input$pp_zoom_to_zero
   do.call(".pp_y_vs_avg_rep", args = list(
     y = y, 
-    colMeans_y_rep = colMeans(y_rep),
+    colMeans_yrep = colMeans(yrep),
     zoom_to_zero = zoom
   ))
 })